[Pictures] Move metadata extraction to metadata folder
