+++ cmake/treedata/common/subdirs.txt
@@ -21,6 +21,7 @@ xbmc/imagefiles                                 imagefiles
 xbmc/messaging                                  messaging
 xbmc/messaging/helpers                          messagingHelpers
 xbmc/pictures                                   pictures
+xbmc/pictures/metadata                          pictures_metadata
 xbmc/platform                                   platform
 xbmc/playlists                                  playlists
 xbmc/powermanagement                            powermanagement
