+++ xbmc/pictures/metadata/Exif.h
