+++ xbmc/pictures/metadata/ImageMetadata.h
