+++ xbmc/pictures/metadata/ImageMetadataParser.cpp
