+++ xbmc/pictures/metadata/ImageMetadataParser.h
