+++ xbmc/pictures/metadata/Iptc.h
